Parallel process runs, has bug that crashes on second loop however. GDAL still broken.

+++ PREACT/PREACTcore/Source/Visualization/FireDomainVisualizer.cs
@@ -16,7 +16,7 @@ namespace PREACT.Visualization
         //LCP
         public abstract void SetAndDisplayLCP(LCPData lcpData, LcpViewMode lcpViewMode = LcpViewMode.FuelModel);
         public abstract void SetLCPViewMode(LcpViewMode lcpViewMode);
-        public abstract void DisplayTriggerBuffer(int[,] data);        
+        public abstract void DisplayTriggerBuffer(float[,] data);        
         public abstract void SetVisibility(bool visible);
         public abstract bool ToggleVisibility();
         public abstract bool IsDataPlaneActive();
